Replaced nrlib by gaussianfft as module for sim og gauss fields in 2D and 3D.
Replaced np.int with np.int32
Removed unused files and functions
Added __init__.py in directory for unit tests to ensure module roxar is imported before gaussianfft

+++ aps/algorithms/defineFaciesProbTrend.py
@@ -265,7 +265,7 @@ class DefineFaciesProb(BaseDefineFacies):
         minimum = np.min(facies_real_values)
         maximum = np.max(facies_real_values)
 
-        prob_index = np.zeros(maximum + 1, np.int)
+        prob_index = np.zeros(maximum + 1, np.int32)
         for i in range(maximum):
             prob_index[i] = -1
         for i in range(len(facies_names_in_real)):
